issue #309
git-svn-id: http://tradelink.googlecode.com/svn/trunk@1314 ccc6a032-7b47-0410-9fd8-1f29c74ce1d2

+++ TradeLinkCommon/TikConstants.cs
@@ -10,7 +10,7 @@ namespace TradeLink.Common
         public const string DOT_EXT = ".TIK";
         public const string WILDCARD_EXT = "*.TIK";
 
-        public const System.Int16 VERSION = 1;
+        public const int VERSION = 1;
 
         // file field identifiers
         public const byte Version = 0; // two bytes of version follow
